refactor(lockfile): extract acquire/release into shared src/lib/lockfile.ts
The PID-based lockfile helpers were private to claude-native-teams.ts and
unusable by any other module that needs safe concurrent writes. Promote them
to a shared module with named exports — preserves the existing semantics
(5s timeout, 50ms poll with jitter, stale-PID detection via kill(pid, 0),
force-acquire with warning after timeout) so claude-native-teams.test.ts
passes unchanged.

Prepares the ground for agent-yaml.ts (next commit) which needs the same
lock semantics for atomic writes to agent.yaml.
